+++ carbon/components/pages/TokenDetailsTab/index.tsx
@@ -1,10 +1,10 @@
+import PoolVolumeOverTimeCard from "components/cards/tokenDetails/PoolVolumeOverTimeCard";
 import TokenDistributionOfMethodologiesCard from "components/cards/tokenDetails/TokenDistributionOfMethodologiesCard";
 import TokenDistributionOfProjectsCard from "components/cards/tokenDetails/TokenDistributionOfProjectsCard";
 import TokenDistributionOfVintageCard from "components/cards/tokenDetails/TokenDistributionOfVintageCard";
 import TokenOriginsCard from "components/cards/tokenDetails/TokenOriginsCard";
 import TokenPoolBreakdownCard from "components/cards/tokenDetails/TokenPoolBreakdownCard";
 import TokenStateOfDigitalCarbonCard from "components/cards/tokenDetails/TokenStateOfDigitalCarbonCard";
-import PoolVolumeOverTimeCard from "components/cards/tokenDetails/TokenVolumeDepositedOverTimeCard";
 import TokenVolumeOverTimeCard from "components/cards/tokenDetails/TokenVolumeOverTimeCard";
 import { TokenDetailsProps } from "components/cards/tokenDetails/helpers";
 import layout from "theme/layout.module.scss";
